refactor(Purchase): 구매 가격 관리 역할을 위한 Purchase 모델 추가
